requirements.txt
Django==5.0.1
django-environ==0.11.2
supabase==2.3.0
python-dotenv==1.0.0
Pillow==10.2.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0

